requirements.txt
asgiref==3.4.1
Django==3.2.6
django-ranged-response==0.2.0
django-simple-captcha
packaging
Pillow==8.3.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
